Fix reverse video for T200 by removing M100 POKE
Replaced subroutine at 8100 with RV$ and NV$.

Previously the code was filled with

    GOSUB 8100: PRINT "Something in reverse video": GOSUB 8100

However, the subroutine at 8100 was using a Model 100 specific POKE to
toggle reverse video. The cleaner way to do this (which also takes less
bytes) is to create escape sequence strings to enter reverse video
(RV$) and return to normal video (NV$)

    RV$=CHR$(27)+"p": NV$=CHR$(27)+"q"

Usage is simple:

    PRINT RV$ "Something in reverse video" NV$

This method will work on the Model 100/102, Tandy 200 and NEC PC-8300.