Use M100 builtin cursor instead of faking one
The blinking underscore cursor was very hard to see. Enabling the
standard cursor looks better and should work on the M100, T200, and
NEC 8201, 8300. This is done by simply printing the escape sequence
"Esc P".

Note that this required adding a semicolon after many of the existing
PRINT statements so that the flashing cursor makes sense. For example,
after the word "LOADING" when the game first starts.

A new subroutine at 7600 has been added that reprints the current
temporary word and positions the cursor at the end to indicate that
the user can keep typing.

Also note: there is a bug in the program where it prints the most
recently entered word onto the next line after a person hits ENTER.
To compensate, I've set the subroutine in 7600 to clear the field
before redrawing it. This is silly, but works for now.