Add wordlist from 2021 (plus some filler words)
The wordle wordlist starts on June 19, 2021, the 170th day of the
year. That meant I needed to come up with 169 new words for this to
work with M100LE.

I used SCOWL to find common five letter words which were not yet in
any of the wordle lists. It was tough to find words that weren't
plurals, past tense, too obscure, or offensive. I erred on the side of
obscurity, but I believe the words are still reasonably fun to play.

Note to self: if I need to do this again, these are the command lines
I built up:

1. Make sure I didn't accidentally duplicate any words in WL2021.DO:

    cat WL202?.DO | tr '[a-z]' '[A-Z]' | sort | uniq -c  | grep -v 1

2. Put possible new words into a file called "newwords":

    cat WL202?.DO | tr '[a-z]' '[A-Z]' | sort | unix2dos > bar;
    egrep -h '^[a-z]{5}$' /usr/share/dict/scowl/english-words.{10,35} |
        tr '[a-z]' '[A-Z]' | egrep -v 'S|ED$' | sort | unix2dos > foo;
    comm -23 foo bar > newwords;
    wc -l newwords