Use Random Access of RAM Directory to quickly read wordlist
Defaults to using WL20xx.CO for compressed wordlists created by
CMPRSS.DO. However, there is code in place so that if WF$ in line 65
is set to end in .DO instead of .CO, it should be able to handle the
standard ASCII document format. (Untested.)

Eventually, this ought to look for both files and use either.
(Probably should look for .CO first since that will be more common.)