test speed of REDRAW WORD routines
Various portable ways of drawing the word typed so far, followed by
underscores, and then positioning the cursor on the first underscore.
Use 'goto 7500' to test the routine GOSUB'd from 7540.
(try 7600, 7700, 7770, 7800, 7850)

Results from 100 iterations

SECONDS		METHOD
 6		?TM$UL$(I):  REM UL$(3) = "___^H^H^H"
 6		PRINT @ position, "_____"; PRINT @ position, TM$
 7		GOSUB 4200: L=5-LEN(TM$):?TM$;STRING$(L,"_");STRING$(L,CHR$(8))
 9		PRINT VT52 positioning "_____" VT52 positioning TM$
10		GOSUB 4200: ?"_____": GOSUB 4200; ?TM$
16		String concatenation for backspaces: A$=A$+LT$, looping twice.

The method using UL$(I) is fast and use portable VT52 cursor movement,
so there should be no problem with it. However, the alternative of
using PRINT @ or LOCATE depending upon which machine was detected is
just as fast, takes up less string space, and may take up less program
space. (Maybe).

Another alternative is to not call 7600 at all and instead add an
Underscore when hitting BKSP and remove one when typing a letter.
(There is no reason the whole word has to be updated.)