RAND social
When playing a random day, the SOCIAL display should replace the month
and day with the "RAND( 178)" instead of replacing the wordlist
(WL2022.CO).

Remove most of the code that tested different methods of printing to
evaluate speed. Left in 7800 which is an alternative to 7600, which
uses PRINT @ / LOCATE instead of VT52 cursor movement.