Fix 8350 MID$(WF$, 5) = RIGHT$(AD$, 2) ;  European date
Replaced with WF$="WL20"+YY$+".CO"

A simple fix, but that entailed having YY$ be a two digit year, which
caused a whole cascade of fixes such that this commit ended up also
fixing the DATE$ problem for machines that use the European format
(YY/MM/DD) instead of American (MM/DD/YY).

Also fixed a minor bug where the year chosen initially was getting
reset when Random is picked. Now, it uses the same YY$ that was
previously selected. It is of note because I removed "CLEAR" as it was
resetting YY$. That required moving the DIM and READ statements to
before line 10 and resetting the WN variable so it doesn't think the
player has won immediately. There may be other bugs lurking from
uninitialized variables that have not surfaced yet.

In the future, it would make sense to do initialization in three
phases:

1. Once per RUN setup: DIM and READ.

2. Variables that should be initialized when playing a new game,
unless the player is playing a RANDOM word (which sets DY and AD$).

3. Variables that should be reinitialized every time the game is
played, whether the player chose AGAIN and entered a date or RANDOM.