Obviate need for LHS MID$() by making SO$ a 2-D array
On the NEC portables, MID$() cannot be on the Left-Hand Side of an
assignment ('='). So, instead of having an array of strings, SO$(i),
which sets individual characters using MID$, now SO$ is a two
dimensional array (guess number, character number).

This makes printing longer because it must be done in a loop to show
all the characters, but it actually makes some of the code easier to
follow as it doesn't have all the extra MID$() wrappers.