Do not set MID$() for alphabet board
Previously, A1$ and A2$ stored the alphabet board and were updated
by writing to MID$(). However, that is not allowed on the NEC
portables, so the new method is to ignore A1$ and A2$ and draw
directly onto the screen for each letter that has changed.

This seems to be faster, or at the least it gives a nice twinkle
effect as the letters are updated.

To do this, the meaning of H$ has been changed so that it is an array
of symbols corresponding to the letters in the guessed word TM$. each
symbol is the character that should be drawn on the alphabet.
("." wrong letter, "?", wrong place, "*" right letter & place).

However, I will likely change H$ back into holding the HINT to be
printed, so that SO$ can instead hold the symbols to be printed.
This will make the implementation of the "[S]OCIAL" routine easier as
it will not have to check for letters and replace them with asterisks.