Fix unmarked letters bug
Renaming the variable H$ to SO$ caused a minor bug when I got
confused and thought I coud "optimize" by using HI$. As Knuth said,
"Premature optimization is the root of all evil."

This bug was subtle, but could be seen with the word "TAUNT" and the
guess "UNTIL". The hint should be "???..", but was showing as ".??..".