Remove trailing ^Z
I had included it at the end of the file thinking it wouldn't harm
anything, and it probably doesn't, but I realized it does mean that
my BASIC tokenizer program would need to be able to detect a bare
^Z and know not to encode it. But it would have to allow a ^Z when
within quotes or a REM statement.

While that isn't too hard for lex to do, it does seem silly to
implement such a thing, so I'll just remove the ^Z from this file.

The main reason I added it was so that I could simply toss the file
into the serial port and have my Tandy 200 recognize when the end of
file. (`cat M100LEl.DO > /dev/ttyUSB0`). But, I do not need to do that
as I have a script now for sending files ("sendtotandy") that
automatically adds in the ^Z.