Send a BASIC program to a TRS-80 Model 100/102/200 via RS232 serial
Use this by hooking up the Model 100/102/200 ("Model T") to a computer
running Unix with an RS232 null serial cable. (Make sure your Unix
box's serial port has "On Chip Software Handshaking" or you will have
lots of errors.)

On the Model T, run this in the BASIC interpreter:

    LOAD "COM:98N1ENN"

On the Unix side, run this script (sendtomodelt).

As each line is sent to the Model T, it will be tokenized and stored
into memory. When the Model T is done, hit enter on the Unix side to
close the serial port and quit the script.

This saves a significant amount of RAM compared to transferring a
BASIC program as a text file and then tokenizing it all at once. In
particular, M100LE.DO has now gotten so large that the latter method
no longer works.