Show ? for the wrong place letters, even if right place elsewhere
Previously this program would show a period "." (incorrect letter) for
any specific letter (say "X") which was used more than once in the
guess if it was also used in the right place. For example, if the word
was XEROX and the guess was EXLAX, previously the program would
incorrectly show a hint of "?...*", when it should be "??..*".

	XEROX	The secret word
	EXLAX	User's guess
	?...*  <- incorrect, old hint
	??..*  <- new behavior

Note that, like the official Wordle, M100LE will show "right letter,
wrong place" whether or not the actual word has two of those letters.
For example,

	XEROX	The secret word
	FEMME	User's guess
	.*..?  <- New hint behavior