Add possible MacOS support /dev/cu.* (untested)
On BSD systems, /dev/cu is the serial port, so now that will be
automatically detected. Or rather, /dev/cu.usbmodem11745 (or some such
silliness) will be detected, as that is what MacOS's Darwin calls it.

Whether this will actually work or not is an open question. I seem to
recall that BSD's `stty` command used a lowercase -f to specify the
device file to modify. There are probably other differences.