Finish adding support for ELSE, RESTORE
Also implemented -c flag for adding a column that counts the number of
different lines that jump to a specific destination. I'm not sure if
this is actually useful, except as an amusing curiousity.

For example,

    $ adjunct/jumpdestinations -c | cut -c1-10 | sort -n | uniq -c
	 31       1
	  3       2
	  1       3
	  1      26

This says that there were 35 destinations in the M100LE program and 31
of those lines (88%) were jumped to from exactly one other line of code.

There were three lines that were jumped to from two places:

    $ adjunct/jumpdestinations -c | awk '$1 == 2'
	  2    GOTO  7010
	  2   GOSUB  7600
	  2    GOTO 10000

And only one line that was jumped to from three places:

    $ adjunct/jumpdestinations -c | awk '$1 == 3'
	  3    GOTO  8900

And then... there was one line which was jumped to from 26 places:

    $ adjunct/jumpdestinations -c | awk '$1 > 3'
	 26   GOSUB  4200

What does all that mean? Not much. It implies that subroutines, other
than line 4200, were not used for sharing common code as much as for
encapsulating concepts to keep the main flow of the program clear.

(The routine at 4200 is the one used to position the cursor
using VT52 escape sequences.)