Workaround NEC's BSAVE exiting program, NEC file size
I do not know if this is a problem on a real NEC portable, but on
Virtual T's emulator, the NEC 8201's BSAVE command exits to "Direct
Mode" when it finishes. To workaround, I've put the BSAVE in a
subroutine, so the user can type "RETURN" to resume the program where
it left off.

Also, it appears BSAVE has a shorter header (5 bytes instead of 6?) or
something because the filesize is one byte shorter than on the Tandy
computers. As a workaround, I've simply increased the size of the data
file by one on the NECs.