Remove WL$ on error if NK (needs kill) flag is set
WL$ is the .CO file that is being written to. When this program hits
an unexpected state, it should automatically delete the .CO file
instead of leaving behind some half-done and confusing file.

New variable, NK, is set to 1 when the file has been created and needs
to be killed on error.