Github Flavored Markdown is a tad silly, so let's be even sillier
For some reason, the <details> tag is not indenting the content.

There does not appear to be a good solution, so I've fallen back on a
kludge from the early days of the World Wide Web: use <ul>...</ul>
(unsorted list with no actual items) to cause an indent.

It's not great because eventually GitHub may fix their CSS and then
our documentation would be double indented. On the other hand, this
makes it much more readable now and it probably won't be awful later
if we forget to remove the <ul> tags.