Remove .BA.K85 and .BA.M10
BASIC Tokenization for the Olivetti M10 and Kyocera Kyotronic-85
appears to be identical to Tandy BASIC tokenization. That leaves NEC's
N82 BASIC as being the only one that is different.

Other than for NEC portables, while it is still possible for a program
to only run on one specific platform (for example, calling routines in
ROM), it will not be because the BASIC file was saved in binary (.BA)
instead of ASCII (.DO) format. Since M100LE was careful to not use any
machine specific commands, we can share a single .BA file for Tandy,
Olivetti, and Kyocera computers.