Kyocera Kyotronic-85 has the SAVEM bug
As with NEC and M10, K85 requires workaround for SAVEM exiting to
"direct mode". Fix by typing "RETURN" for the user using the keyboard
buffer.

Unlike the NEC and M10, the K85's buffer is stored in every other
byte in memory. (The Tandy portables are the same as the K85, but
those don't have the SAVEM bug.)

Blob regenerated by Makefile